worker/app/requirements01.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
attrs==25.3.0
azure-ai-documentintelligence==1.0.2
azure-core==1.34.0
azure-servicebus==7.14.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.115.12
ffmpeg-python==0.2.0
frozenlist==1.6.2
future==1.0.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
isodate==0.7.2
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.64
langchain-openai==0.3.21
langchain-text-splitters==0.3.8
langgraph==0.4.8
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.45
marshmallow==3.26.1
multidict==6.4.4
mypy_extensions==1.1.0
numpy==2.3.0
openai==1.84.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
propcache==0.3.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydub==0.25.1
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SpeechRecognition==3.14.3
SQLAlchemy==2.0.41
starlette==0.46.2
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
